Control.Distributed.Dataset

Documentation

data Dataset a

Represents a partitioned multiset that can be transformed in a distributed fashion.

Transformations

dMap :: (StaticSerialise a, StaticSerialise b) => Closure (a -> b) -> Dataset a -> Dataset b

Returns a new Dataset that contains the result of applying the given function to each element.

dFilter :: StaticSerialise a => Closure (a -> Bool) -> Dataset a -> Dataset a

Returns a new Dataset that only contains elements where the given function returns true.

dConcatMap :: (StaticSerialise a, StaticSerialise b) => Closure (a -> [b]) -> Dataset a -> Dataset b

Returns a new Dataset by first applying a function to all elements of this Dataset, and then flattening the results.

dGroupedAggr

Arguments

:: (StaticHashable k, StaticSerialise k, StaticSerialise b)

=> Int
Target number of partitions

-> Closure (a -> k)
Grouping key

-> Aggr a b

-> Dataset a

-> Dataset (k, b)

Apply an aggregation to all rows sharing the same key.

dDistinct :: StaticHashable a => Int -> Dataset a -> Dataset a

Removes a new dataset with duplicate rows removed.

dDistinctBy

Arguments

:: StaticHashable b

=> Int
Target number of partitions

-> Closure (a -> b)

-> Dataset a

-> Dataset a

Removes a new dataset with rows with the duplicate keys removed.

Low-level transformations

dCoalesce :: Typeable a => Int -> Dataset a -> Dataset a

Coalesce partitions together to get the specified number of partitions.

dPipe :: (StaticSerialise a, StaticSerialise b) => Closure (ConduitT a b (ResourceT IO) ()) -> Dataset a -> Dataset b

Transforms a Dataset by passing every partition through the given Conduit.

dPartition :: (StaticSerialise a, StaticHashable k) => Int -> Closure (a -> k) -> Dataset a -> Dataset a

Re-partition the dataset using the given function so that the items with the same k will end up in the same partition.

Aggregations

module Control.Distributed.Dataset.Aggr

Execution

dAggr :: (StaticSerialise a, StaticSerialise b) => Aggr a b -> Dataset a -> DD b

Apply an aggregation to all items on a Dataset, and fetch the result.

dFetch :: StaticSerialise a => Dataset a -> DD (ConduitT () a (ResourceT IO) ())

Returns a Conduit to fetch the results lazily to the driver.

dToList :: StaticSerialise a => Dataset a -> DD [a]

Fetches the complete dataset as a list.

data DD a

runDD :: Backend -> ShuffleStore -> DD a -> IO a

runDDWith :: LogLevel -> Backend -> ShuffleStore -> DD a -> IO a

data Backend

Backend is responsible for running your functions in a remote environment.

See:

localProcessBackend

distributed-dataset-aws

data ShuffleStore

Provides a way to store intermediate temporary data.

See:

withLocalTmpShuffleStore

distributed-dataset-aws

Creating datasets

data Partition a

Represents some amount of data which to be transformed on a single executor.

mkPartition :: Typeable a => Closure (ConduitT () a (ResourceT IO) ()) -> Partition a

Create a Partition given a Source conduit.

dExternal :: StaticSerialise a => [Partition a] -> Dataset a

Create a dataset from given Partition's.

This is how every Dataset is created initially.

Class

class (Typeable a, Serialise a) => StaticSerialise a where

Methods

staticSerialise :: Closure (Dict (Typeable a, Serialise a))

class (Typeable a, Eq a, Hashable a) => StaticHashable a where

Methods

staticHashable :: Closure (Dict (Typeable a, Eq a, Hashable a))

Re-exports

initDistributedFork :: IO ()

On distributed-fork, we run the same binary both in your machine (called "driver") and in the remote environment (called "executor"). In order for the program to act according to where it is, you should call this function as the first thing in your main:

main = do
  initDistributedFork
  ...

liftIO :: MonadIO m => IO a -> m a

Lift a computation from the IO monad.

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

>>> 5 & (+1) & show
"6"

Since: base-4.8.0.0

Closure

data Closure a

Type of serializable closures. Abstractly speaking, a closure is a code reference paired together with an environment. A serializable closure includes a shareable code reference (i.e. a Closure). Closures can be serialized only if all expressions captured in the environment are serializable.

cap :: Typeable a => Closure (a -> b) -> Closure a -> Closure b

Closure application. Note that Closure is not a functor, let alone an applicative functor, even if it too has a meaningful notion of application.

cpure :: Closure (Dict (Serializable a)) -> a -> Closure a

A closure can be created from any serializable value. cpure corresponds to Control.Applicative's pure, but restricted to lifting serializable values only.

data Dict a where

Values of type Dict p capture a dictionary for a constraint of type p.

e.g.

Dict :: Dict (Eq Int)

captures a dictionary that proves we have an:

instance Eq 'Int

Pattern matching on the Dict constructor will bring this instance into scope.

Constructors

Dict :: forall a. a => Dict a