Control.Distributed.Fork.Backend

Description

You only need this module if you want to create a new backend for distributed-dataset.

See localProcessBackend for a minimal example.

Writing a Backend.

newtype Backend

Backend is responsible for running your functions in a remote environment.

See:

localProcessBackend

distributed-dataset-aws

Constructors

Backend

Fields

bExecute :: ByteString -> BackendM ByteString

Should run the current binary in the target environment, put the given string as standard input and return the executables answer on the standard output.

data BackendM a

BackendM is essentially IO, but also has the ability to report the status of the executor.

argExecutorMode :: String

We switch to executor mode only when argv[1] == argExecutorMode.

Reporting status

data ExecutorFinalStatus a

Constructors

ExecutorFailed Text

ExecutorSucceeded a

data ExecutorStatus a

Constructors

ExecutorPending ExecutorPendingStatus

ExecutorFinished (ExecutorFinalStatus a)

data ExecutorPendingStatus

Constructors

ExecutorWaiting (Maybe Text)

ExecutorSubmitted (Maybe Text)

ExecutorStarted (Maybe Text)

waiting :: BackendM ()

waitingDesc :: Text -> BackendM ()

submitted :: BackendM ()

submittedDesc :: Text -> BackendM ()

started :: BackendM ()

startedDesc :: Text -> BackendM ()

Utils

throttledBackend :: MonadIO m => Int -> Backend -> m Backend

Re-exports

liftIO :: MonadIO m => IO a -> m a

Lift a computation from the IO monad.

getExecutablePath :: IO FilePath

Returns the absolute pathname of the current executable.

Note that for scripts and interactive sessions, this is the path to the interpreter (e.g. ghci.)

Since base 4.11.0.0, getExecutablePath resolves symlinks on Windows. If an executable is launched through a symlink, getExecutablePath returns the absolute path of the original executable.

Since: base-4.6.0.0