Control.Distributed.Fork

Description

This module provides a common interface for offloading an IO action to remote executors.

Documentation

fork :: MonadIO m => Backend -> Closure (Dict (Serializable a)) -> Closure (IO a) -> m (Handle a)

Asynchronously executes the given function using the Backend and returns an Handle.

initDistributedFork :: IO ()

On distributed-fork, we run the same binary both in your machine (called "driver") and in the remote environment (called "executor"). In order for the program to act according to where it is, you should call this function as the first thing in your main:

data Backend

Backend is responsible for running your functions in a remote environment.

Should run the current binary in the target environment, put the given string as standard input and return the executables answer on the standard output. | BackendM is essentially IO, but also has the ability to report the status of the executor.

Handle

data Handle a

Result of a fork is an Handle where you can await a result.

await :: (MonadIO m, MonadThrow m) => Handle a -> m a

Blocks until the Handle completes.

ExecutorStatus

pollHandle :: Handle a -> STM (ExecutorStatus a)

Get the current status of given Handle.

data ExecutorStatus a

Constructors

ExecutorPending ExecutorPendingStatus

ExecutorFinished (ExecutorFinalStatus a)

data ExecutorPendingStatus

Constructors

ExecutorWaiting (Maybe Text)

ExecutorSubmitted (Maybe Text)

ExecutorStarted (Maybe Text)

data ExecutorFinalStatus a

Constructors

ExecutorFailed Text

ExecutorSucceeded a

Exceptions

newtype ExecutorFailedException

Constructors

ExecutorFailedException Text

Re-exports

type Serializable a = (Binary a, Typeable a)

Values that can be sent across the network.

data Closure a

Type of serializable closures. Abstractly speaking, a closure is a code reference paired together with an environment. A serializable closure includes a shareable code reference (i.e. a StaticPtr). Closures can be serialized only if all expressions captured in the environment are serializable.

cap :: Typeable a => Closure (a -> b) -> Closure a -> Closure b

Closure application. Note that Closure is not a functor, let alone an applicative functor, even if it too has a meaningful notion of application.

cpure :: Closure (Dict (Serializable a)) -> a -> Closure a

A closure can be created from any serializable value. cpure corresponds to Control.Applicative's pure, but restricted to lifting serializable values only.

data Dict a where

Values of type Dict p capture a dictionary for a constraint of type p.

e.g.

Dict :: Dict (Eq Int)

captures a dictionary that proves we have an:

instance Eq 'Int

Pattern matching on the Dict constructor will bring this instance into scope.

Constructors

Dict :: forall a. a => Dict a