Control.Distributed.Fork.Backend

Description

You only need this module if you want to create a new backend for distributed-dataset.

Writing a Backend.

newtype Backend

Backend is responsible for running your functions in a remote environment.

Should run the current binary in the target environment, put the given string as standard input and return the executables answer on the standard output. | BackendM is essentially IO, but also has the ability to report the status of the executor.

Constructors

Backend

Fields

bExecute :: ByteString -> BackendM ByteString

data BackendM a

argExecutorMode :: String

We switch to executor mode only when argv[1] == argExecutorMode.

Reporting status

data ExecutorFinalStatus a

Constructors

ExecutorFailed Text

ExecutorSucceeded a

data ExecutorStatus a

Constructors

ExecutorPending ExecutorPendingStatus

ExecutorFinished (ExecutorFinalStatus a)

data ExecutorPendingStatus

Constructors

ExecutorWaiting (Maybe Text)

ExecutorSubmitted (Maybe Text)

ExecutorStarted (Maybe Text)

waiting :: BackendM ()

waitingDesc :: Text -> BackendM ()

submitted :: BackendM ()

submittedDesc :: Text -> BackendM ()

started :: BackendM ()

startedDesc :: Text -> BackendM ()

Utils

throttledBackend :: MonadIO m => Int -> Backend -> m Backend

runBackend :: Closure (Dict (Serializable i)) -> Closure (IO i) -> Backend -> IO (Handle i)

Given an IO action and a static proof that the result is Serializable, this function runs the action using the Backend in a separate thread and returns a TVar holding the ExecutorStatus.

toBackendStdin :: Closure (Dict (Serializable a)) -> Closure (IO a) -> ByteString

fromBackendStdout :: Binary a => ByteString -> ExecutorFinalStatus a

Re-exports

liftIO :: MonadIO m => IO a -> m a

Lift a computation from the IO monad.

getExecutablePath :: IO FilePath

Returns the absolute pathname of the current executable.

Note that for scripts and interactive sessions, this is the path to the interpreter (e.g. ghci.)

Since base 4.11.0.0, getExecutablePath resolves symlinks on Windows. If an executable is launched through a symlink, getExecutablePath returns the absolute path of the original executable.

Since: base-4.6.0.0